git push

+++ src/main/java/com/repository/FraisCycleRepository.java
@@ -3,6 +3,6 @@ package com.repository;
 import com.entities.FraisCycle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FraisCycleRepository extends JpaRepository<FraisCycle, Integer> {
+public interface FraisCycleRepository extends JpaRepository<FraisCycle, Long> {
 
 }
