+++ src/main/java/com/repository/InspectionRepository.java
@@ -3,6 +3,6 @@ package com.repository;
 import com.entities.Inspection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InspectionRepository extends JpaRepository<Inspection, Integer> {
+public interface InspectionRepository extends JpaRepository<Inspection, Long> {
 
 }
